+++ src/main/java/dev/fucksable/FuckSableMixinConfigPlugin.java
@@ -10,7 +10,6 @@ import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 
 import java.io.InputStream;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;